+++ core/src/sdk/examples/simple/src/ClientTest.cpp
@@ -143,9 +143,9 @@ ClientTest::Test(const std::string& address, const std::string& port) {
 
     milvus_sdk::Utils::Sleep(3);
     {  // search vectors
-        std::vector<std::string> partiton_tags;
+        std::vector<std::string> partition_tags;
         milvus::TopKQueryResult topk_query_result;
-        milvus_sdk::Utils::DoSearch(conn, TABLE_NAME, partiton_tags, TOP_K, NPROBE, search_record_array,
+        milvus_sdk::Utils::DoSearch(conn, TABLE_NAME, partition_tags, TOP_K, NPROBE, search_record_array,
                                     topk_query_result);
     }
 
@@ -169,9 +169,9 @@ ClientTest::Test(const std::string& address, const std::string& port) {
     }
 
     {  // search vectors
-        std::vector<std::string> partiton_tags;
+        std::vector<std::string> partition_tags;
         milvus::TopKQueryResult topk_query_result;
-        milvus_sdk::Utils::DoSearch(conn, TABLE_NAME, partiton_tags, TOP_K, NPROBE, search_record_array,
+        milvus_sdk::Utils::DoSearch(conn, TABLE_NAME, partition_tags, TOP_K, NPROBE, search_record_array,
                                     topk_query_result);
     }